✨ added: refactored images & icons, constants.dart, Login UI
