+++ pubspec.yaml
@@ -62,7 +62,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
-    - assets/icons
+    - assets/icons/
 
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
